A message reader Application will allow you to save your important messages on a single tap.
The main aim of MESSAGE READER APPLICATION, named VIMessages (Very
Important Messages), is to bridge this gap of security between the
user’s ease of use and their personal data. This application features a
sign-up and login activity, which makes it secure. Also if the user
forgets the pin or wants to change it, it can be accomplished easily.
Also an option of account deletion is available.

The key feature of this application is the user-friendly interface to
view the inbox messages and save the personal ones with just one tap.
The saved messages are saved in SQLite database, which can be viewed
easily in another fragment. Also all the saved messages can be deleted
with just one tap in case of an emergency. An additional feature of
sending an SMS through the application is also available.